Move create_pipeline.py into a separate module
